pyproject.toml
[project]
name = "csc591lab3"
version = "0.1.0"
description = "CSC591 Lab3: Building a Transformer Language Model from Scratch"
readme = "README.md"
requires-python = ">=3.10"
authors = [
    { name = "Gavin Gong", email = "zgong6@ncsu.edu" }
]
keywords = ["transformer", "language-model", "pytorch", "deep-learning", "education"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Education",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "torch>=2.0.0",
    "numpy>=1.24.0",
    "matplotlib>=3.7.0",
    "datasets>=2.18.0",
    "tokenizers>=0.15.0",
    "tqdm>=4.65.0",
    "pyyaml>=6.0",
    "einops>=0.7.0",
    "ipykernel>=7.1.0",
    "ipython>=8.38.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",
]
wandb = [
    "wandb>=0.16.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
markers = [
    "part1: Part 1 component and composition tests",
    "part2: Part 2+ staged tests",
    "part2_tokenizer: Part 2 tokenizer-focused tests",
    "part3_training: Training-loop and checkpoint tests",
    "part4_generation: Text-generation tests",
    "stretch: Optional advanced features for stronger graduate-level submissions",
    "integration: end-to-end integration coverage across multiple modules",
]

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]
include = '\.pyi?$'

[tool.ruff]
line-length = 100
target-version = "py310"
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
